CB-16504 Outbound only load balancer option
Adds a new load balancer type, OUTBOUND. An OUTBOUND load balancer is a public load balancer that
only does egress, and does not have any rules for routing ingress traffic. This is used in Azure
clusters when public IPs are disabled. With the Standard load balancer, any VMs behind a private LB
will not have a path for public egress. So in that scenario, an outbound only load balancer will
also be created solely to handle public egress.

This also changes the default Azure load balancer SKU from BASIC to STANDARD.

Tested by setting up a cluster with only the private LB and verifying the VMs behind it had no path
to the internet, and then setting up a cluster with a private LB and a second outbound only LB and
verifying the VMs could now do public egress over the internet.